+++ Data/WrathClassic/ExportedTooltipData.lua
@@ -1,4 +1,4 @@
--- [[ Exported at 2022-12-04 09-38-52 ]] --
+-- [[ Exported at 2022-12-11 22-08-41 ]] --
 -- [[ This code is automatically generated as an export from ]] --
 -- [[ an SQLite database and is not meant for manual edit. ]] --
 
@@ -27,12 +27,14 @@ function exportedTooltipData.Load(d)
     t[8] = addon.L["Already eaten for achievement"];
     t[4] = addon.L["Already killed for achievement"];
     t[10] = addon.L["Already drank for achievement"];
+    t[14] = addon.L["Already petted for achievement"];
     t[12] = addon.L["Already defeated for achievement"];
     t[1] = addon.L["Needs /love for achievement"];
     t[5] = addon.L["Needs to be cooked for achievement"];
     t[7] = addon.L["Needs to be eaten for achievement"];
     t[3] = addon.L["Needs to be killed for achievement"];
     t[9] = addon.L["Needs to be drunk for achievement"];
+    t[13] = addon.L["Needs to be petted for achievement"];
     t[11] = addon.L["Needs to be defeated for achievement"];
 
     AddD(d, 2442, tdt.Unit, 1206, 1, t[1], t[2]); -- "Cow"